refactor(mobile-filter): rename mocks, rename story file
